add code

+++ SWP391_Recipe_Organizer_BE.API/ViewModel/IngredientVM.cs
@@ -5,9 +5,9 @@
         public string? IngredientId { get; set; }
         public string? IngredientName { get; set; }
         public string? Measure { get; set; }
-        public int? Carbohydrate { get; set; } = 0;
-        public int? Protein { get; set; } = 0;
-        public int? Fat { get; set; } = 0;
-        public int? Calories { get; set; } = 0;
+        public float? Carbohydrate { get; set; } = 0;
+        public float? Protein { get; set; } = 0;
+        public float? Fat { get; set; } = 0;
+        public float? Calories { get; set; } = 0;
     }
 }
