+++ SWP391_Recipe_Organizer_BE.Repo/EntityModel/Ingredient.cs
@@ -13,10 +13,10 @@ namespace SWP391_Recipe_Organizer_BE.Repo.EntityModel
         public string IngredientId { get; set; } = null!;
         public string? IngredientName { get; set; }
         public string? Measure { get; set; }
-        public int? Carbohydrate { get; set; }
-        public int? Protein { get; set; }
-        public int? Fat { get; set; }
-        public int? Calories { get; set; }
+        public double? Carbohydrate { get; set; }
+        public double? Protein { get; set; }
+        public double? Fat { get; set; }
+        public double? Calories { get; set; }
         public bool? IsDelete { get; set; }
 
         public virtual ICollection<IngredientOfRecipe> IngredientOfRecipes { get; set; }
